Initial attempt at generalizing the project for #57.
Wrapped most files in an `init()`  function that will attach
their various method to the new `ASTLib` instance.
