+++ lib/equiv.js
@@ -1,5 +1,8 @@
+/* global require, exports */
 var assert = require("assert");
-var types = require("../main");
+
+exports.init = function(types) {
+
 var getFieldNames = types.getFieldNames;
 var getFieldValue = types.getFieldValue;
 var isArray = types.builtInTypes.array;
@@ -7,6 +10,7 @@ var isObject = types.builtInTypes.object;
 var isDate = types.builtInTypes.Date;
 var isRegExp = types.builtInTypes.RegExp;
 var hasOwn = Object.prototype.hasOwnProperty;
+var typesP = Object.getPrototypeOf(types);
 
 function astNodesAreEquivalent(a, b, problemPath) {
     if (isArray.check(problemPath)) {
@@ -102,6 +106,7 @@ function arraysAreEquivalent(a, b, problemPath) {
 }
 
 function objectsAreEquivalent(a, b, problemPath) {
+    var i, name;
     isObject.assert(a);
     if (!isObject.check(b)) {
         return false;
@@ -122,8 +127,8 @@ function objectsAreEquivalent(a, b, problemPath) {
     var bNameCount = bNames.length;
 
     if (aNameCount === bNameCount) {
-        for (var i = 0; i < aNameCount; ++i) {
-            var name = aNames[i];
+        for (i = 0; i < aNameCount; ++i) {
+            name = aNames[i];
             var aChild = getFieldValue(a, name);
             var bChild = getFieldValue(b, name);
 
@@ -175,4 +180,6 @@ function objectsAreEquivalent(a, b, problemPath) {
     return false;
 }
 
-module.exports = astNodesAreEquivalent;
+typesP.astNodesAreEquivalent = astNodesAreEquivalent;
+
+};
