+++ lib/node-path.js
@@ -1,10 +1,14 @@
+/* global exports, require */
 var assert = require("assert");
-var types = require("./types");
+
+exports.init = function(types) {
+
 var n = types.namedTypes;
 var isNumber = types.builtInTypes.number;
 var isArray = types.builtInTypes.array;
-var Path = require("./path");
-var Scope = require("./scope");
+var Path = types.Path;
+var Scope = types.Scope;
+var typesP = Object.getPrototypeOf(types);
 
 function NodePath(value, parentPath, name) {
     assert.ok(this instanceof NodePath);
@@ -374,4 +378,6 @@ function firstInStatement(path) {
     return true;
 }
 
-module.exports = NodePath;
+typesP.NodePath = NodePath;
+
+};
