+++ lib/path-visitor.js
@@ -1,11 +1,15 @@
+/* global require, exports */
 var assert = require("assert");
-var types = require("./types");
-var NodePath = require("./node-path");
+
+exports.init = function(types) {
+types.registerDefs("../def/core");
+var NodePath = types.NodePath;
 var Node = types.namedTypes.Node;
 var isArray = types.builtInTypes.array;
 var isObject = types.builtInTypes.object;
 var isFunction = types.builtInTypes.function;
 var hasOwn = Object.prototype.hasOwnProperty;
+var typesP = Object.getPrototypeOf(types);
 var undefined;
 
 function PathVisitor() {
@@ -46,7 +50,7 @@ PathVisitor.fromMethodsObject = function fromMethodsObject(methods) {
 
     if (!isObject.check(methods)) {
         // An empty visitor?
-        return new PathVisitor;
+        return new PathVisitor();
     }
 
     function Visitor() {
@@ -63,7 +67,7 @@ PathVisitor.fromMethodsObject = function fromMethodsObject(methods) {
     isFunction.assert(Visitor.fromMethodsObject);
     isFunction.assert(Visitor.visit);
 
-    return new Visitor;
+    return new Visitor();
 };
 
 function extend(target, source) {
@@ -231,7 +235,7 @@ function invokeVisitorMethod(methodName) {
 
 sharedContextProtoMethods.traverse =
 function traverse(path, newVisitor) {
-    assert.ok(this instanceof this.Context);
+    assert.ok(this instanceof this.visitor.Context);
     assert.ok(path instanceof NodePath);
     assert.ok(this.currentPath instanceof NodePath);
 
@@ -242,4 +246,6 @@ function traverse(path, newVisitor) {
     ));
 };
 
-module.exports = PathVisitor;
+typesP.PathVisitor = PathVisitor;
+typesP.visit = PathVisitor.visit;
+};
