+++ lib/path.js
@@ -1,12 +1,16 @@
+/* global require, exports */
 var assert = require("assert");
 var Op = Object.prototype;
 var hasOwn = Op.hasOwnProperty;
-var types = require("./types");
+
+exports.init = function(types) {
+
 var isArray = types.builtInTypes.array;
 var isNumber = types.builtInTypes.number;
 var Ap = Array.prototype;
 var slice = Ap.slice;
 var map = Ap.map;
+var typesP = Object.getPrototypeOf(types);
 
 function Path(value, parentPath, name) {
     assert.ok(this instanceof Path);
@@ -80,7 +84,7 @@ Pp.each = function each(callback, context) {
     var i = 0;
 
     // Collect all the original child paths before invoking the callback.
-    for (var i = 0; i < len; ++i) {
+    for (i = 0; i < len; ++i) {
         if (hasOwn.call(this.value, i)) {
             childPaths[i] = this.get(i);
         }
@@ -190,7 +194,7 @@ Pp.unshift = function unshift(node) {
 
 Pp.push = function push(node) {
     isArray.assert(this.value);
-    delete getChildCache(this).length
+    delete getChildCache(this).length;
     return this.value.push.apply(this.value, arguments);
 };
 
@@ -277,6 +281,7 @@ function repairRelationshipWithParent(path) {
 }
 
 Pp.replace = function replace(replacement) {
+    var i;
     var results = [];
     var parentValue = this.parentPath.value;
     var parentCache = getChildCache(this.parentPath);
@@ -345,4 +350,6 @@ Pp.replace = function replace(replacement) {
     return results;
 };
 
-module.exports = Path;
+typesP.Path = Path;
+
+};
