+++ lib/scope.js
@@ -1,15 +1,18 @@
+/* global exports, require */
 var assert = require("assert");
-var types = require("./types");
+
+exports.init = function(types) {
 var Type = types.Type;
 var namedTypes = types.namedTypes;
 var Node = namedTypes.Node;
 var isArray = types.builtInTypes.array;
 var hasOwn = Object.prototype.hasOwnProperty;
 var b = types.builders;
+var typesP = Object.getPrototypeOf(types);
 
 function Scope(path, parentScope) {
     assert.ok(this instanceof Scope);
-    assert.ok(path instanceof require("./node-path"));
+    assert.ok(path instanceof types.NodePath);
     ScopeType.assert(path.value);
 
     var depth;
@@ -241,4 +244,6 @@ Sp.getGlobalScope = function() {
     return scope;
 };
 
-module.exports = Scope;
+typesP.Scope = Scope;
+
+};
