+++ lib/shared.js
@@ -1,26 +1,32 @@
-var types = require("../lib/types");
+/* global require, exports */
+
+exports.init = function(types) {
 var Type = types.Type;
 var builtin = types.builtInTypes;
 var isNumber = builtin.number;
+var typesP = Object.getPrototypeOf(types);
+
+typesP.shared = {};
+var sharedP = Object.getPrototypeOf(typesP.shared);
 
 // An example of constructing a new type with arbitrary constraints from
 // an existing type.
-exports.geq = function(than) {
+function geq(than) {
     return new Type(function(value) {
         return isNumber.check(value) && value >= than;
     }, isNumber + " >= " + than);
-};
+}
 
 // Default value-returning functions that may optionally be passed as a
 // third argument to Def.prototype.field.
-exports.defaults = {
+var defaults = {
     // Functions were used because (among other reasons) that's the most
     // elegant way to allow for the emptyArray one always to give a new
     // array instance.
-    "null": function() { return null },
-    "emptyArray": function() { return [] },
-    "false": function() { return false },
-    "true": function() { return true },
+    "null": function() { return null; },
+    "emptyArray": function() { return []; },
+    "false": function() { return false; },
+    "true": function() { return true; },
     "undefined": function() {}
 };
 
@@ -32,10 +38,27 @@ var naiveIsPrimitive = Type.or(
     builtin.undefined
 );
 
-exports.isPrimitive = new Type(function(value) {
+var isPrimitive = new Type(function(value) {
     if (value === null)
         return true;
     var type = typeof value;
     return !(type === "object" ||
              type === "function");
 }, naiveIsPrimitive.toString());
+
+Object.defineProperties(sharedP, {
+    geq : {
+        enumerable: true,
+        value: geq
+    },
+    defaults: {
+        enumerable: true,
+        value: defaults
+    },
+    isPrimitive: {
+        enumerable: true,
+        value: isPrimitive
+    }
+});
+
+};
