+++ lib/traverse.js
@@ -1,5 +1,9 @@
-var visit = require("./path-visitor").visit;
+/* global exports, console */
+exports.init = function(types) {
+
+var visit = types.visit;
 var warnedAboutDeprecation = false;
+var typesP = Object.getPrototypeOf(types);
 
 function traverseWithFullPathInfo(node, callback) {
     if (!warnedAboutDeprecation) {
@@ -24,4 +28,6 @@ function traverseWithFullPathInfo(node, callback) {
 }
 
 traverseWithFullPathInfo.fast = traverseWithFullPathInfo;
-module.exports = traverseWithFullPathInfo;
+typesP.traverseWithFullPathInfo = traverseWithFullPathInfo;
+typesP.traverse = traverseWithFullPathInfo;
+};
