+++ lib/types.js
@@ -1,3 +1,5 @@
+/* global require, exports */
+
 var assert = require("assert");
 var Ap = Array.prototype;
 var slice = Ap.slice;
@@ -8,6 +10,9 @@ var objToStr = Op.toString;
 var funObjStr = objToStr.call(function(){});
 var strObjStr = objToStr.call("");
 var hasOwn = Op.hasOwnProperty;
+
+exports.init = function(types) {
+var typesP = Object.getPrototypeOf(types);
 
 // A type is an object with a .check method that takes a value and returns
 // true or false according to whether the value matches the type.
@@ -44,7 +49,7 @@ var Tp = Type.prototype;
 
 // Throughout this file we use Object.defineProperty to prevent
 // redefinition of exported properties.
-exports.Type = Type;
+types.Type = Type;
 
 // Like .check, except that failure triggers an AssertionError.
 Tp.assert = function(value, deep) {
@@ -80,20 +85,17 @@ Tp.toString = function() {
     return name + " type";
 };
 
-var builtInTypes = {};
-exports.builtInTypes = builtInTypes;
-
 function defBuiltInType(example, name) {
     var objStr = objToStr.call(example);
 
-    Object.defineProperty(builtInTypes, name, {
+    Object.defineProperty(types.builtInTypes, name, {
         enumerable: true,
         value: new Type(function(value) {
             return objToStr.call(value) === objStr;
         }, name)
     });
 
-    return builtInTypes[name];
+    return types.builtInTypes[name];
 }
 
 // These types check the underlying [[Class]] attribute of the given
@@ -105,7 +107,7 @@ var isFunction = defBuiltInType(function(){}, "function");
 var isArray = defBuiltInType([], "array");
 var isObject = defBuiltInType({}, "object");
 var isRegExp = defBuiltInType(/./, "RegExp");
-var isDate = defBuiltInType(new Date, "Date");
+var isDate = defBuiltInType(new Date(), "Date");
 var isNumber = defBuiltInType(3, "number");
 var isBoolean = defBuiltInType(true, "boolean");
 var isNull = defBuiltInType(null, "null");
@@ -245,14 +247,11 @@ Fp.getValue = function(obj) {
 // type d.type by calling methods such as d.bases, d.build, and d.field.
 Type.def = function(typeName) {
     isString.assert(typeName);
-    return hasOwn.call(defCache, typeName)
-        ? defCache[typeName]
-        : defCache[typeName] = new Def(typeName);
+    return hasOwn.call(types.defCache, typeName)
+        ? types.defCache[typeName]
+        : types.defCache[typeName] = new Def(typeName);
 };
 
-// In order to return the same Def instance every time Type.def is called
-// with a particular name, those instances need to be stored in a cache.
-var defCache = Object.create(null);
 
 function Def(typeName) {
     var self = this;
@@ -281,8 +280,8 @@ Def.fromValue = function(value) {
     if (value && typeof value === "object") {
         var type = value.type;
         if (typeof type === "string" &&
-            hasOwn.call(defCache, type)) {
-            var d = defCache[type];
+            hasOwn.call(types.defCache, type)) {
+            var d = types.defCache[type];
             if (d.finalized) {
                 return d;
             }
@@ -306,9 +305,9 @@ Dp.isSupertypeOf = function(that) {
 
 // Note that the list returned by this function is a copy of the internal
 // supertypeList, *without* the typeName itself as the first element.
-exports.getSupertypeNames = function(typeName) {
-    assert.ok(hasOwn.call(defCache, typeName));
-    var d = defCache[typeName];
+typesP.getSupertypeNames = function(typeName) {
+    assert.ok(hasOwn.call(types.defCache, typeName));
+    var d = types.defCache[typeName];
     assert.strictEqual(d.finalized, true);
     return d.supertypeList.slice(1);
 };
@@ -316,12 +315,12 @@ exports.getSupertypeNames = function(typeName) {
 // Returns an object mapping from every known type in the defCache to the
 // most specific supertype whose name is an own property of the candidates
 // object.
-exports.computeSupertypeLookupTable = function(candidates) {
+typesP.computeSupertypeLookupTable = function(candidates) {
     var table = {};
 
-    for (var typeName in defCache) {
-        if (hasOwn.call(defCache, typeName)) {
-            var d = defCache[typeName];
+    for (var typeName in types.defCache) {
+        if (hasOwn.call(types.defCache, typeName)) {
+            var d = types.defCache[typeName];
             assert.strictEqual(d.finalized, true);
             for (var i = 0; i < d.supertypeList.length; ++i) {
                 var superTypeName = d.supertypeList[i];
@@ -421,25 +420,19 @@ Dp.bases = function() {
 // False by default until .build(...) is called on an instance.
 Object.defineProperty(Dp, "buildable", { value: false });
 
-var builders = {};
-exports.builders = builders;
-
-// This object is used as prototype for any node created by a builder.
-var nodePrototype = {};
-
 // Call this function to define a new method to be shared by all AST
 // nodes. The replaced method (if any) is returned for easy wrapping.
-exports.defineMethod = function(name, func) {
-    var old = nodePrototype[name];
+typesP.defineMethod = function(name, func) {
+    var old = types.nodePrototype[name];
 
     // Pass undefined as func to delete nodePrototype[name].
     if (isUndefined.check(func)) {
-        delete nodePrototype[name];
+        delete types.nodePrototype[name];
 
     } else {
         isFunction.assert(func);
 
-        Object.defineProperty(nodePrototype, name, {
+        Object.defineProperty(types.nodePrototype, name, {
             enumerable: true, // For discoverability.
             configurable: true, // For delete proto[name].
             value: func
@@ -477,18 +470,18 @@ Dp.build = function(/* param1, param2, ... */) {
     // Every buildable type will have its "type" field filled in
     // automatically. This includes types that are not subtypes of Node,
     // like SourceLocation, but that seems harmless (TODO?).
-    self.field("type", self.typeName, function() { return self.typeName });
+    self.field("type", self.typeName, function() { return self.typeName; });
 
     // Override Dp.buildable for this Def instance.
     Object.defineProperty(self, "buildable", { value: true });
 
-    Object.defineProperty(builders, getBuilderName(self.typeName), {
+    Object.defineProperty(types.builders, getBuilderName(self.typeName), {
         enumerable: true,
 
         value: function() {
             var args = arguments;
             var argc = args.length;
-            var built = Object.create(nodePrototype);
+            var built = Object.create(types.nodePrototype);
 
             assert.ok(
                 self.finalized,
@@ -577,9 +570,6 @@ Dp.field = function(name, type, defaultFn, hidden) {
     return this; // For chaining.
 };
 
-var namedTypes = {};
-exports.namedTypes = namedTypes;
-
 // Like Object.keys, but aware of what fields each AST type should have.
 function getFieldNames(object) {
     var d = Def.fromValue(object);
@@ -595,7 +585,7 @@ function getFieldNames(object) {
 
     return Object.keys(object);
 }
-exports.getFieldNames = getFieldNames;
+typesP.getFieldNames = getFieldNames;
 
 // Get the value of an object property, taking object.type and default
 // functions into account.
@@ -610,13 +600,13 @@ function getFieldValue(object, fieldName) {
 
     return object[fieldName];
 }
-exports.getFieldValue = getFieldValue;
+typesP.getFieldValue = getFieldValue;
 
 // Iterate over all defined fields of an object, including those missing
 // or undefined, passing each field name and effective value (as returned
 // by getFieldValue) to the callback. If the object has no corresponding
 // Def, the callback will never be called.
-exports.eachField = function(object, callback, context) {
+typesP.eachField = function(object, callback, context) {
     getFieldNames(object).forEach(function(name) {
         callback.call(this, name, getFieldValue(object, name));
     }, context);
@@ -626,7 +616,7 @@ exports.eachField = function(object, callback, context) {
 // callback returns a truthy value. Like Array.prototype.some, the final
 // result is either true or false to indicates whether the callback
 // returned true for any element or not.
-exports.someField = function(object, callback, context) {
+typesP.someField = function(object, callback, context) {
     return getFieldNames(object).some(function(name) {
         return callback.call(this, name, getFieldValue(object, name));
     }, context);
@@ -644,7 +634,7 @@ Dp.finalize = function() {
         var allSupertypes = this.allSupertypes;
 
         this.baseNames.forEach(function(name) {
-            var def = defCache[name];
+            var def = types.defCache[name];
             def.finalize();
             extend(allFields, def.allFields);
             extend(allSupertypes, def.allSupertypes);
@@ -663,7 +653,7 @@ Dp.finalize = function() {
         }
 
         // Types are exported only once they have been finalized.
-        Object.defineProperty(namedTypes, this.typeName, {
+        Object.defineProperty(types.namedTypes, this.typeName, {
             enumerable: true,
             value: this.type
         });
@@ -683,7 +673,7 @@ function populateSupertypeList(typeName, list) {
 
     for (var pos = 0; pos < list.length; ++pos) {
         typeName = list[pos];
-        var d = defCache[typeName];
+        var d = types.defCache[typeName];
         assert.strictEqual(d.finalized, true);
 
         // If we saw typeName earlier in the breadth-first traversal,
@@ -715,10 +705,12 @@ function extend(into, from) {
     });
 
     return into;
+}
+
+types.finalize = function() {
+    Object.keys(types.defCache).forEach(function(name) {
+        types.defCache[name].finalize();
+    });
 };
 
-exports.finalize = function() {
-    Object.keys(defCache).forEach(function(name) {
-        defCache[name].finalize();
-    });
 };
